remove reducer file, use index instead
