+++ src/core/permissions/reducers/index.js
@@ -2,8 +2,34 @@
  * @flow
  */
 
-import reducer from './reducer';
+import { Map, fromJS } from 'immutable';
 
-export { default as gatherOrganizationPermissionsReducer } from './gatherOrganizationPermissionsReducer';
+import gatherOrganizationPermissionsReducer from './gatherOrganizationPermissionsReducer';
 
-export default reducer;
+import { RESET_REQUEST_STATE } from '../../redux/actions';
+import { ACES, RS_INITIAL_STATE } from '../../redux/constants';
+import { resetRequestStateReducer } from '../../redux/reducers';
+import {
+  GATHER_ORGANIZATION_PERMISSIONS,
+  gatherOrganizationPermissions,
+} from '../actions';
+
+const INITIAL_STATE :Map = fromJS({
+  // actions
+  [GATHER_ORGANIZATION_PERMISSIONS]: RS_INITIAL_STATE,
+  // data
+  [ACES]: Map(),
+});
+
+export default function reducer(state :Map = INITIAL_STATE, action :Object) {
+
+  if (action.type === RESET_REQUEST_STATE) {
+    return resetRequestStateReducer(state, action);
+  }
+
+  if (action.type === gatherOrganizationPermissions.case(action.type)) {
+    return gatherOrganizationPermissionsReducer(state, action);
+  }
+
+  return state;
+}
